Adds unit tests for ShoppingService and ShoppingItemDto

+++ src/main/java/mx/kinich49/itemtracker/Constants.java
@@ -6,5 +6,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class Constants {
 
-    public static final int PRICE_SCALE = 100;
+    public static final double PRICE_SCALE = 100;
 }
